// Training/Frontend/src/app/interface/Model/ModelTraining.class.ts
export class ModelTraining {
    Id: string;
    PartnerId: string;
    SubPartnerId: string;
    TrainingCenterId: string;
    TrainerId: string;
    Type:string;
    TrainingTypeId: string;
    TrainingCategoryId: string;
    CityId: string;
    AreaId: string;
    StartDate: Date;
    EndDate: Date;
    days: string[];
    IsOnline: boolean;
}